refactor: SessionManager.kt 패키지 이동
