+++ android/app/src/main/java/com/on/turip/ui/compose/invitation/InvitationEntryViewModel.kt
@@ -5,11 +5,11 @@ import androidx.lifecycle.viewModelScope
 import com.on.turip.core.result.ErrorType
 import com.on.turip.core.result.onFailure
 import com.on.turip.core.result.onSuccess
-import com.on.turip.core.session.SessionManager
 import com.on.turip.core.session.SessionState
 import com.on.turip.domain.invitation.InvitationTokenParser
 import com.on.turip.domain.invitation.usecase.DetermineInvitationEntryRouteUseCase
 import com.on.turip.domain.invitation.usecase.model.InvitationEntryResult
+import com.on.turip.domain.session.SessionManager
 import com.on.turip.domain.turip.repository.TuripRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.channels.Channel
